almost working

+++ include/off_line_knowledge.h
@@ -14,15 +14,15 @@ struct Off_line_knowledge_key {
 
 struct Off_line_knowledge_map : json_cpp::Json_object{
     Off_line_knowledge_map(const cell_world::Graph &);
-    void save_reward(const cell_world::Cell &, const cell_world::Cell &, const cell_world::Cell &, double);
-    const cell_world::Cell &get_destination(const cell_world::Cell &, const cell_world::Cell &, double);
-    const cell_world::Cell &get_destination(const cell_world::Model_public_state &, double);
+    void save_reward(const cell_world::Cell &, const cell_world::Cell &, const cell_world::Cell &, float);
+    const cell_world::Cell &get_destination(const cell_world::Cell &, const cell_world::Cell &, float);
+    const cell_world::Cell &get_destination(const cell_world::Model_public_state &, float);
     cell_world::Move get_move(const cell_world::Model_public_state &);
-    json_cpp::Json_vector<double> &get_rewards(const cell_world::Cell &, const cell_world::Cell &);
-    json_cpp::Json_vector<double> &get_rewards(const cell_world::Model_public_state &);
+    json_cpp::Json_vector<float> &get_rewards(const cell_world::Cell &, const cell_world::Cell &);
+    json_cpp::Json_vector<float> &get_rewards(const cell_world::Model_public_state &);
 
     const cell_world::Graph &_graph;
-    json_cpp::Json_vector<json_cpp::Json_vector<json_cpp::Json_vector<double>>> rewards;
+    json_cpp::Json_vector<json_cpp::Json_vector<json_cpp::Json_vector<float>>> rewards;
     json_cpp::Json_vector<json_cpp::Json_vector<json_cpp::Json_vector<unsigned int>>> counters;
     Json_object_members({
         Add_member(rewards);
